Combine search queries with union
Performance is inconsitent and often bad without limiting the span
of the search headline. It might need to be limited further. The
count query (caused by pagination) is often slow too. This should
be investigated to see if a simpler query could provide the count.

Most fields are removed from the template, as we're working towards
a simpler page anyway.