Rework search page
Several things needed attention, such as broken pagination, code
duplication and general complexity. The tag soup "worked", but was
obviously not tennable.

This tries to solve all of that, gets rid of almost all custom CSS
classes in favour of bootstrap, and moves to more semantic HTML and
a few usability and accessibility improvements, e.g.
 - blockquote, section, etc.
 - larger checkboxes
 - Show message when there are no results
 - aria-label here and there

This moves the filters to the right, which hopefully ensures more
attention on the results. The second search input is removed in
favour of links that assist users on small screens to go to the
"other" part of the page. The assumption is that enough context is
visible on larger screens.

And the page is down from 220kB to under 50kB.