Rework page error handling
The change is smaller than the diff looks. The error block is now
a section with role="alert". The block is always hidden before a
request, so that a duplicate error (second 404, or second network
error) will also be announced. This allowed some simplification of
the handling of `hidden`. We escape strings from translation to
ensure they are valid JavaScript strings. The rest is cosmetic.