build: pass ALLOWED_HOSTS env var to app
This variable allows you to set which IPs the app may be accessed from. This is useful to make it accessible when testing on a LAN network with another device (e.g. a mobile phone). The list of allowed hosts is passed as a space-separated string in the environment variable (e.g. `ALLOWED_HOSTS="127.0.0.1 localhost 192.168.100.25"`)