fix(a11y): clarity when start/end dates are missing
Before this change, if the start or end dates were missing for a project, the screenreader would simply read "End Date:" or "Start Date:" and trail off. Now, it will read "End Date: Ongoing" and "Start Date: unspecified". For sighted users, the text is slightly greyed-out to indicate that the end date is not _really_ there (as the project is ongoing).