chore(ci): don't run push & PR on PR branches
This change prevents duplicate runs from occurring when opening a pull request
on GitHub. Before, both the 'push' and 'pull_request' hooks would trigger,
meaning that any push to a branch with an open PR would run the CI pipeline
twice. By filtering the 'push' hook to 'main' and 'develop', we avoid this.