feat: allow users to edit the document fulltext
This allows users that have the suitable permission to edit the fulltext field
of Documents. This field is used when searching, so it allows users to remove
frontmatter from PDFs which are irrelevant, or paste a plaintext version for a
Google Doc link, for example. The field is only shown to those with the relevant
permission to avoid confusing people who may not understand what it is for.
When a PDF is edited for a Document, unless the fulltext is also edited at the
same time (within the same form submission), it will take priority and
overwrite the old fulltext (even if edited prior).