test(selenium): run tests across multiple {browser, js} combinations
By passing the browser type and whether JS is enabled via environment variables,
we can run the Selenium frontend tests on all combinations of targetted
browsers and JS/no-JS in the CI pipeline. For now, we are not targetting Safari
for this, as the GitHub Actions MacOS workers don't suppose `services`, which
we use to run the postgres database. This will be left for a follow-up (see
issue #145).