Avoid long lists of documents
This hopefully makes rendering and browsing the languages and
subjects pages a bit simpler, and refactors a reused bit. We can
consider doing the same for projects.

The custom CSS is to work correctly without Bootstrap's JS on browsers
that support the [open] selector (should be almost 100% supported now),
and to fix a webkit/safari bug (visible with the GNOME browser).