Fix+tweak styling of htmx-indicator
`animation` now correctly has a delay of 1s, which was the idea all
along. Alpha is added to the static background so that it would not
obscure underlying content before the animation starts. This way
no indicator is shown if the response arives fast enough. Alpha is
increased for the indicator to compensate for any loss of contrast
with the white background now gone.